Rename CSRF error file name after code review
 On branch feature/sec-115-scrf-restricted-word-web
 Your branch is up to date with 'origin/feature/sec-115-scrf-restricted-word-web'.

 Changes to be committed:
	modified:   src/app.ts
	renamed:    src/middleware/csrfErrorMiddleware.ts -> src/middleware/createCsrfErrorMiddleware.ts
